+++ lib/sail/engine.rb
@@ -6,11 +6,22 @@ module Sail
   # after initialize hooks
   class Engine < ::Rails::Engine
     require "jquery-rails" if Rails::VERSION::MAJOR < 5
+    require "sprockets/railtie"
     isolate_namespace Sail
 
     config.generators do |g|
       g.test_framework :rspec
     end
+
+    config.middleware.use ActionDispatch::Flash
+    config.middleware.use ActionDispatch::Cookies
+    config.middleware.use ActionDispatch::Session::CookieStore
+    config.middleware.use ActionDispatch::ContentSecurityPolicy::Middleware if defined?(ActionDispatch::ContentSecurityPolicy)
+    config.middleware.use Rack::MethodOverride
+    config.middleware.use Rails::Rack::Logger
+    config.middleware.use Rack::Head
+    config.middleware.use Rack::ConditionalGet
+    config.middleware.use Rack::ETag
 
     initializer "sail.assets.precompile" do |app|
       app.config.assets.precompile += %w[sail/refresh.svg sail/sort.svg sail/angle-left.svg